Move subscriptions methods to a separate dir inside lib
