+++ lib/subscription/sendVerificationMail.ts
